requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
decorator==4.4.2
distlib==0.3.2
Django==3.2.6
django-choices==1.7.2
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.21.1
openapi-codec==1.3.2
Pillow==8.3.1
platformdirs==2.2.0
playsound==1.3.0
proglog==0.1.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
simplejson==3.17.5
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
tk==0.1.0
tqdm==4.62.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
zope.interface==5.4.0